Modularize header and mainCointainer and they styles files
