+++ .gitignore
@@ -11,7 +11,7 @@ node_modules
 dist
 dist-ssr
 *.local
-*d.ts
+
 
 # Editor directories and files
 .vscode/*
