+++ package.json
@@ -4,6 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
+    "build:dev": "tsc --w",
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
