+++ src/header/styles.css
@@ -1,0 +1,3 @@
+.title {
+  font-size: 3rem;
+}
