+++ src/index.ts
@@ -1,19 +1,3 @@
-import createMainHeader from "./header.ts";
+import renderMainContainer from "./mainContainer/mainContainer";
 import "./style.css";
-
-const renderMainContainer = () => {
-  const root = document.querySelector(".root");
-
-  const mainContainer = document.createElement("div");
-  mainContainer.classList.add("main-container");
-
-  const header = createMainHeader("Monumentos del Mundo");
-  if (!root) {
-    throw Error("Element not found");
-  }
-
-  mainContainer.appendChild(header);
-  root.appendChild(mainContainer);
-};
-
 renderMainContainer();
