+++ src/mainContainer/styles.css
@@ -1,0 +1,7 @@
+.main-container {
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  max-width: 1500px;
+  padding: 16px;
+}
