+++ src/style.css
@@ -1,3 +1,7 @@
+:root {
+  --main-background-color: #0077d4;
+}
+
 *,
 ::after,
 ::before {
@@ -7,21 +11,9 @@
 body {
   margin: 0;
   font-family: Arial, Helvetica, sans-serif;
-  background-color: #198290;
+  background-color: var(--main-background-color);
 }
 
 h1 {
   margin: 0;
-}
-
-.main-container {
-  margin: 0 auto;
-  display: flex;
-  justify-content: center;
-  max-width: 90vw;
-  padding: 16px;
-}
-
-.title {
-  font-size: 3rem;
 }
